Moved custom UltiSnips snippets directory
According to UltiSnips manpage, UltiSnips checks `runtimepath` from nvim
to check for directories with snippets.

`g:UltiSnipsSnippetDirectories=["UltiSnips"]` means, that UltiSnips will
check for directories named `UltiSnips` for paths in `runtimepath`.

Currently in runtimepath: ~/.config/nvim, so the directory with the
custom snippets need to be moved to this directory.